fix(ci): build dashboard before cargo, drop tag trigger from deploy-docs, clippy cleanups + fmt

+++ crates/ua-cli/src/util/time.rs
@@ -20,7 +20,7 @@ pub fn iso8601_now() -> String {
 pub fn epoch_to_ymdhms(secs: i64) -> (i64, u32, u32, u32, u32, u32) {
     let z = secs.div_euclid(86_400) + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64;
+    let doe = z - era * 146_097;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146_096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
